fix: writer WAL directory not respecting WRITER_WAL_DIR environment variable
The writer service was using a clap default value for wal_dir but wasn't
configured to read from the WRITER_WAL_DIR environment variable. This
caused the WAL files to be created in the wrong location when using
./scripts/run-dev.sh.

Added env = "WRITER_WAL_DIR" to the clap argument to properly read from
the environment variable set by the development script.

Fixes issue where writer WAL was created in /.wal instead of /.data/wal/writer